requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq==0.4.1
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
requests==2.31.0
websockets==12.0
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1
